Update: split structure to frontend, backend and common
