+++ rust-warp-jwt-example/Cargo.toml
@@ -1,15 +1,9 @@
-[package]
-name = "rust-warp-jwt-example"
-version = "0.1.0"
-edition = "2021"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+members = [
+    "backend",
+    "frontend",
 
-[dependencies]
-jsonwebtoken = "9"
-serde = { version = "1.0", features = ["derive"]}
-serde_json = "1.0"
-thiserror = "1.0"
-tokio = { version = "1", features = ["full"] }
-warp = "0.3"
-chrono = "0.4.35"
+    #Internal
+    "common"
+]
