+++ rust-warp-jwt-example/backend/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "backend"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+jsonwebtoken = "9"
+serde = { version = "1.0", features = ["derive"]}
+serde_json = "1.0"
+thiserror = "1.0"
+tokio = { version = "1", features = ["full"] }
+warp = "0.3"
+chrono = "0.4.35"
+common = { version = "0.1.0", path = "../common" }
