+++ rust-warp-jwt-example/common/src/lib.rs
@@ -1,0 +1,7 @@
+#[derive(Clone)]
+pub struct User{
+    pub uid: String,
+    pub email: String,
+    pub pw: String,
+    pub role: String,
+}
