+++ rust-warp-jwt-example/frontend/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "frontend"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+yew = "0.21"
+wasm-bindgen = "0.2.67"
+serde_json = "1"
+serde = {version = "=1.0", features = ["derive"] }
+anyhow = "1"
+yew-router = "0.18.0"
+common = { version = "0.1.0", path = "../common" }
