+++ rust-warp-jwt-example/frontend/index.html
@@ -1,0 +1,61 @@
+<html>
+  <head>
+    <title>Rust Fullstack Example</title>
+    <style>
+        body {
+            font-size: 14px;
+            font-family: sans-serif;
+        }
+        a {
+            text-decoration: none;
+            color: #339;
+        }
+        a:hover {
+            text-decoration: none;
+            color: #33f;
+        }
+        .app {
+            background-color: #efefef;
+            margin: 100px 25% 25% 25%;
+            width: 50%;
+            padding: 10px;
+        }
+        .app .nav {
+            text-align: center;
+            font-size: 16px;
+            font-weight: bold;
+        }
+        .app .refresh {
+            text-align: center;
+            margin: 10px 0 10px 0;
+        }
+        .list-item {
+            margin: 2px;
+            padding: 5px;
+            background-color: #cfc;
+        }
+        .pet {
+            margin-top: 10px;
+        }
+        .completed {
+            text-decoration: line-through;
+            background-color: #dedede;
+        }
+        .detail {
+            font-size: 16px;
+        }
+        .detail h1 {
+            font-size: 24px;
+        }
+        .detail .id {
+            color: #999;
+        }
+        .detail .completed {
+            color: #3f3;
+        }
+        .detail .not-completed {
+            color: #f33;
+        }
+    </style>
+  </head>
+</html>
